Increased UI usability and added functionality to detail of book

+++ BookstarApp/app/src/main/java/com/example/sergio/bookstarapp/room/BookEntity.kt
@@ -9,7 +9,7 @@ data class BookEntity(
   val title: String,
   val author: String,
   val isFavorite: Boolean,
-  val coverId: Int,
+  val coverId: Long,
   val hasFullText: Boolean,
   val editionCount: Int,
   val firstPublishYear: Int
